Fix Responses provider transport after alpha-test rebase
