+++ crates/app/src/provider/request_failover_runtime.rs
@@ -51,6 +51,7 @@ where
                         try_next_model,
                         reason,
                         snapshot,
+                        ..
                     } = model_error;
                     record_provider_failover_audit_event(
                         kernel_ctx,
